+++ modules/ads_lib/src/test/java/com/google/api/ads/adwords/lib/client/reporting/ReportingConfigurationTest.java
@@ -159,10 +159,6 @@ public class ReportingConfigurationTest {
             // A timeout of 0 is acceptable. This means do not time out.
             .reportDownloadTimeout(0)
             .build();
-    config.validate("v201705");
-    assertTrue("Validation should succeed for v201705", true);
-    config.validate("v201708");
-    assertTrue("Validation should succeed for v201708", true);
     config.validate("v201710");
     assertTrue("Validation should succeed for v201710", true);
     config.validate("v201802");