+++ modules/adwords_appengine/src/main/java/com/google/api/ads/adwords/jaxws/utils/v201705/batchjob/package-info.java
@@ -1,5 +1,0 @@
-@javax.xml.bind.annotation.XmlSchema(
-  namespace = "https://adwords.google.com/api/adwords/cm/v201705",
-  elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED
-)
-package com.google.api.ads.adwords.jaxws.utils.v201705.batchjob;